Refactor: 브랜드 좋아요 토글
