+++ src/main/java/org/example/sivillage/brand/application/BrandService.java
@@ -15,6 +15,7 @@ public interface BrandService {
      * 2. getBrandIdList 브랜드 ID 목록 조회
      * 3. getBrandName 브랜드 정보 조회
      * 4. getBrandLike 브랜드 좋아요 여부 조회
+     * 5. toggleBrandLike 브랜드 좋아요 토글
      */
 
 
@@ -46,4 +47,12 @@ public interface BrandService {
      * return GetBrandLikeResponseDto
      */
     GetBrandLikeResponseDto getBrandLike(Long brandId, String memberUuid);
+
+    /**
+     * 5. toggleBrandLike 브랜드 좋아요 토글
+     * @param brandId 브랜드 ID
+     * @param memberUuid 회원 UUID
+     * return void
+     */
+    void toggleBrandLike(Long brandId, String memberUuid);
 }
